lỗi tùm lum , fix sau

+++ LanAnhComputer.API/Data/Entities/Cart.cs
@@ -4,8 +4,8 @@ namespace LanAnhComputer.API.Data.Entities
 {
     public class Cart
     {
-        public long CartId { get; set; }
-        public long UserId { get; set; }
+        public int CartId { get; set; }
+        public int UserId { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime? UpdatedAt { get; set; }
